Move nuki coordinator to separate module
